Refactored client models
