+++ MyMoney.Client/Interfaces/Api/IBudgetApi.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using MyMoney.Client.Models.DTO;
+using MyMoney.Client.Models.Entity;
 using MyMoney.Client.Models.Request;
 
 namespace MyMoney.Client.Interfaces.Api
