+++ MyMoney.Client/Interfaces/Api/ITransactionApi.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
-using MyMoney.Client.Models.DTO;
+using MyMoney.Client.Models.Common;
+using MyMoney.Client.Models.Entity;
 using MyMoney.Client.Models.Request;
 using MyMoney.Client.Models.Response;
