+++ MyMoney.Client/Models/Entity/BudgetModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MyMoney.Client.Models.DTO
+namespace MyMoney.Client.Models.Entity
 {
     public class BudgetModel
     {
