+++ MyMoney.Common/ViewModels/TransactionListViewModel.cs
@@ -2,7 +2,8 @@
 using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using MyMoney.Client.Models.DTO;
+using MyMoney.Client.Models.Common;
+using MyMoney.Client.Models.Entity;
 using MyMoney.Client.Models.Request;
 using MyMoney.Common.Views;
 using Xamarin.Forms;
